Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -55,7 +55,7 @@ end
 def get_average_age_for_season(data, season)
   contestants = data[season]
   #binding.pry
-  age_sum = contestants.inject(0) {|sum, i| sum + i['age'].to_i}
+  age_sum = contestants.inject(0) {|sum, i| sum + i['age'].to_f}
   age_sum / contestants.length
   
 end
